// model/batchGetAccountsRequest.ts
/**
 * ManageSpace CRM API
 * ManageSpace CRM API Documentation
 *
 * The version of the OpenAPI document: 0.0.11-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class BatchGetAccountsRequest {
    /**
    * A list of UUIDs that represent account ids.
    */
    'accountIds': Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accountIds",
            "baseName": "accountIds",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return BatchGetAccountsRequest.attributeTypeMap;
    }
}